Restore caching on DN normalization and sheet date parsing
