+++ app/api/dn/query.py
@@ -1,0 +1,130 @@
+"""DN record querying endpoints."""
+
+from __future__ import annotations
+
+from datetime import datetime
+from typing import Optional, List
+
+from fastapi import APIRouter, Depends, HTTPException, Query
+from sqlalchemy.orm import Session
+
+from app.core.sync import DN_RE, VALID_STATUS_DESCRIPTION
+from app.crud import list_dn_records, list_dn_records_by_dn_numbers, search_dn_records
+from app.db import get_db
+from app.utils.query import normalize_batch_dn_numbers
+from app.utils.string import normalize_dn
+from app.utils.time import to_gmt7_iso
+
+router = APIRouter(prefix="/api/dn")
+
+
+@router.get("/search")
+def search_dn_records_api(
+    dn_number: Optional[str] = Query(None, description="精确 DN number"),
+    status: Optional[str] = Query(None, description=f"状态过滤，可选: {VALID_STATUS_DESCRIPTION}"),
+    remark: Optional[str] = Query(None, description="备注关键词(模糊)"),
+    has_photo: Optional[bool] = Query(None, description="是否必须带附件 true/false"),
+    date_from: Optional[datetime] = Query(None, description="起始时间(ISO 8601)"),
+    date_to: Optional[datetime] = Query(None, description="结束时间(ISO 8601)"),
+    page: int = Query(1, ge=1),
+    page_size: Optional[int] = Query(None, ge=1, description="每页数量，缺省时返回全部符合条件的数据"),
+    db: Session = Depends(get_db),
+):
+    if dn_number:
+        dn_number = normalize_dn(dn_number)
+        if not DN_RE.fullmatch(dn_number):
+            raise HTTPException(status_code=400, detail=f"Invalid DN number: {dn_number}")
+
+    total, items = search_dn_records(
+        db,
+        dn_number=dn_number,
+        status=status,
+        remark_keyword=remark,
+        has_photo=has_photo,
+        date_from=date_from,
+        date_to=date_to,
+        page=page,
+        page_size=page_size,
+    )
+
+    return {
+        "ok": True,
+        "total": total,
+        "page": page,
+        "page_size": page_size if page_size is not None else len(items),
+        "items": [
+            {
+                "id": it.id,
+                "dn_number": it.dn_number,
+                "status": it.status,
+                "remark": it.remark,
+                "photo_url": it.photo_url,
+                "lng": it.lng,
+                "lat": it.lat,
+                "updated_by": it.updated_by,
+                "created_at": to_gmt7_iso(it.created_at),
+            }
+            for it in items
+        ],
+    }
+
+
+@router.get("/batch")
+def batch_get_dn_records(
+    dn_number: Optional[List[str]] = Query(None, description="重复 dn_number 或逗号分隔"),
+    dnnumber_legacy: Optional[List[str]] = Query(
+        None, alias="dnnumber", description="重复 dn_number 或逗号分隔 (legacy alias)", include_in_schema=False
+    ),
+    page: int = Query(1, ge=1),
+    page_size: int = Query(20, ge=1, le=100),
+    db: Session = Depends(get_db),
+):
+    dn_numbers = normalize_batch_dn_numbers(dn_number, dnnumber_legacy)
+
+    total, items = list_dn_records_by_dn_numbers(db, dn_numbers, page=page, page_size=page_size)
+    return {
+        "ok": True,
+        "total": total,
+        "page": page,
+        "page_size": page_size,
+        "items": [
+            {
+                "id": it.id,
+                "dn_number": it.dn_number,
+                "status": it.status,
+                "remark": it.remark,
+                "photo_url": it.photo_url,
+                "lng": it.lng,
+                "lat": it.lat,
+                "updated_by": it.updated_by,
+                "created_at": to_gmt7_iso(it.created_at),
+            }
+            for it in items
+        ],
+    }
+
+
+@router.get("/{dn_number}")
+def get_dn_records(dn_number: str, db: Session = Depends(get_db)):
+    dn_number = normalize_dn(dn_number)
+    if not DN_RE.fullmatch(dn_number):
+        raise HTTPException(status_code=400, detail="Invalid DN number")
+
+    items = list_dn_records(db, dn_number)
+    return {
+        "ok": True,
+        "items": [
+            {
+                "id": it.id,
+                "dn_number": it.dn_number,
+                "status": it.status,
+                "remark": it.remark,
+                "photo_url": it.photo_url,
+                "lng": it.lng,
+                "lat": it.lat,
+                "updated_by": it.updated_by,
+                "created_at": to_gmt7_iso(it.created_at),
+            }
+            for it in items
+        ],
+    }
