+++ app/api/health.py
@@ -1,0 +1,10 @@
+"""Health endpoint."""
+
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get("/")
+def healthz():
+    return {"ok": True, "message": "You can use admin panel now."}
